# `<locale>`: Double-checked locking for locale::classic