--- a/stl/src/locale0.cpp
+++ b/stl/src/locale0.cpp
@@ -148,20 +148,23 @@ __PURE_APPDOMAIN_GLOBAL locale::id codecvt<unsigned short, char, mbstate_t>::id(
 #endif // hardware
 
 _MRTIMP2_PURE const locale& __CLRCALL_PURE_OR_CDECL locale::classic() { // get reference to "C" locale
+#if !defined(_M_CEE_PURE)
     const auto mem = reinterpret_cast<const intptr_t*>(&locale::_Locimp::_Clocptr);
     intptr_t as_bytes;
 #ifdef _WIN64
 #ifdef _M_ARM
     as_bytes = __ldrexd(_Mem);
-#else
+#else // ^^^ ARM64 / x64 vvv
     as_bytes = __iso_volatile_load64(mem);
-#endif
-#else // ^^^ _WIN64 / !_WIN64 vvv
+#endif // ^^^ x64
+#else // ^^^ 64-bit / 32-bit vvv
     as_bytes = __iso_volatile_load32(mem);
-#endif // ^^^ !_WIN64 ^^^
+#endif // ^^^ 32-bit ^^^
     _Compiler_or_memory_barrier();
     const auto ptr = reinterpret_cast<locale::_Locimp*>(as_bytes);
-    if (ptr == nullptr) {
+    if (ptr == nullptr)
+#endif // !defined(_M_CEE_PURE)
+    {
         _Init();
     }
 
@@ -188,7 +191,10 @@ _MRTIMP2_PURE locale::_Locimp* __CLRCALL_PURE_OR_CDECL locale::_Init(bool _Do_in
         // set classic to match
         ptr->_Incref();
         ::new (&classic_locale) locale(ptr);
-        const auto mem      = reinterpret_cast<volatile intptr_t*>(&locale::_Locimp::_Clocptr);
+#if defined(_M_CEE_PURE)
+        locale::_Locimp::_Clocptr = ptr;
+#else // ^^^ _M_CEE_PURE / !_M_CEE_PURE vvv
+        const auto mem = reinterpret_cast<volatile intptr_t*>(&locale::_Locimp::_Clocptr);
         const auto as_bytes = reinterpret_cast<intptr_t>(ptr);
         _Compiler_or_memory_barrier();
 #ifdef _WIN64
@@ -196,6 +202,7 @@ _MRTIMP2_PURE locale::_Locimp* __CLRCALL_PURE_OR_CDECL locale::_Init(bool _Do_in
 #else
         __iso_volatile_store32(mem, as_bytes);
 #endif
+#endif // ^^^ !defined(_M_CEE_PURE) ^^^
     }
 
     if (_Do_incref) {