Closes #3030.

The original issue suggests double-checked locking of `global_locale` in `locale::_Init`, but I don't think that would work without significantly more work, if it's possible at all. The reason is that the global locale can be modified, while at the same time a default constructed `locale` could be trying to load the global locale and increment it's refcount. This is the kind of lock-free "juggling with knives" that I don't want to even try. The classic locale, on the other hand, is a typical singleton and more amenable to DCL.

ABI alert: `InterlockedMeowPointer` requires their arguments be 32/64 bit aligned, thus the added `alignas`. Is that an ABI issue?