@MattStephanson 2022-08-21
> I don't see a justification to use WinAPI / Intrinsics.
> 
> If you can change the locking tech, you may go with `std::call_once` as out-of-box DCL. Or `std::atomic` if `std::call_once does not suit. You are in a .cpp and don't need to worry about includes and a downgrade standard mode.
> 
> If you need to preserve the variable, you can still use `std::atomic_ref`, but then I doubt if this change is safe at all.

On the original issue, @CaseyCarter noted (via @StephanTLavavej):
> - This needs to be compatible with `/clr:pure` (which is incompatible with atomics),
> - This is injected into the STL's import library, where we recently had trouble with `atomic` appearing (and fixed it by not dragging in `atomic`)