@MattStephanson 2022-08-21
> That's not a good idea to have DCL via CAS. On x86 and x64 it is an expensive operation. Still a win over a mutex, as a mutex would yeld two such ops, but still need to have plain read, if want real DCL gain.

Yes, I'm not a fan of the CAS. I think `<atomic>` uses either `__iso_volatile_meow` or `__ldrexd` and `_Compiler_or_memory_barrier`, which is either `_ReadWriteBarrier` or `__dmb(0xB)`. Do you think that would be better?

https://github.com/microsoft/STL/blob/fef8191adb2cee1daa0399fde7733413cd6d4a98/stl/inc/atomic#L155-L167

https://github.com/microsoft/STL/blob/fef8191adb2cee1daa0399fde7733413cd6d4a98/stl/inc/atomic#L1036-L1045

https://github.com/microsoft/STL/blob/fef8191adb2cee1daa0399fde7733413cd6d4a98/stl/inc/atomic#L97-L107