@AlexGuteniev 2022-08-22
> Yes, I'm not a fan of the CAS. I think `<atomic>` uses either `__iso_volatile_meow` or `__ldrexd` and `_Compiler_or_memory_barrier`, which is either `_ReadWriteBarrier` or `__dmb(0xB)`. Do you think that would be better?

For x86/x64 -- yes, absolutely.
For ARM situation is as follows:
 * Ideally we want `LDAR` here, but we can't now, see #83 (Current ARM atomics are suboptimal due to that and other issues)
 * `InterlockedCompareExchangePointerAcquire` does not compile to true ARM64 CAS, it is LL/SC, so it consists of load, store, a barrier, and a retry loop.
 * I don't have a sense on ARM perf, and cannot tell whether `__dmb(0xB)` + plain load is better or worse, but my guess it is that fake CAS described above is worse than `__dmb(0xB)` + plain load, but the real CAS could have been better

My understanding is that we should primarily optimize for x86 / x64, and don't have much divergence between these and ARM, so should go with a barrier plus a load.