@AlexGuteniev 2022-08-22
> On the original issue, @CaseyCarter noted (via @StephanTLavavej):
> 
> > * This needs to be compatible with `/clr:pure` (which is incompatible with atomics),
> > * This is injected into the STL's import library, where we recently had trouble with `atomic` appearing (and fixed it by not dragging in `atomic`)

I see, but is there a problem with `std::call_once`, or its WinAPI equivalent?