@StephanTLavavej 2022-08-24
Thanks! We're very interested in reviewing this performance enhancement, but because locales are by far the scariest area of the STL, we'd like to target merging this at the beginning of the VS 2022 17.5 Preview 1 cycle (so we have time to react to any bug reports), instead of the end of the 17.4 Preview 3 cycle. Fortunately, this won't require waiting long - the date when changes start flowing into 17.5 Preview 1 is 2022-09-02 (i.e. next week Friday).