@MattStephanson 2022-08-30
> I see, but is there a problem with `std::call_once`, or its WinAPI equivalent?

Not necessarily, but (a) some rough tests show it's about 2x slower on my machine (3.85 ns vs 1.70 ns for copy-pasting the atomic load/store code; always-locking code is 16.7 ns) and (b) there's some business in `xonce2.cpp` with `/ALTERNATENAME` that I don't understand.