@AlexGuteniev 2022-08-30
> it's about 2x slower on my machine (3.85 ns vs 1.70 ns for copy-pasting the atomic load/store code; always-locking code

Well, that's expected. I assumed it to be uniform for x86/ARM, and faster than `InterlockedCompareExchangePointer`, but not faster than plain load due to indirection.