@StephanTLavavej 2022-10-26
FYI @barcharcraz, we've made changes since you approved.

```
------------------------------------------------------------
Benchmark                  Time             CPU   Iterations
------------------------------------------------------------
main:
BM_locale_classic       30.1 ns         29.2 ns     23578947

This PR:
BM_locale_classic       2.89 ns         2.92 ns    235789474
```

Thanks @MattStephanson for this major improvement in a notoriously problematic area! :tada: